Fix access to the userinfo
passport-cloudfoundry fails to load the user profile from uaa.
uaa refuses to serve the /userinfo requests to passport-cloudfoundry
it redirects to the login view as the access token is not set in the
headers.

node-oauth will not add the accesstoken to the headers
when it performs a GET request unless the method
useAuthorizationHeaderforGET has been called.

This is the case for the versions of node-oauth installed by
the current version of passport-oauth (1.1.15)
and the earlier versions used by passport-cloudfoundry (0.1.x)